+++ Sources/SwiftExtras/Extensions/View/View+Error.swift
@@ -24,7 +24,7 @@ extension View {
     ///
     ///     var body: some View {
     ///         VStack {
-    ///             Text("Hello, World!")
+    ///             Text("Hello World!")
     ///                 .padding()
     ///         }
     ///         .task {