+++ Sources/SwiftExtras/Modifiers/PulsatingEffect.swift
@@ -43,7 +43,7 @@ public extension View {
 #if DEBUG
 @available(iOS 17, macOS 14, tvOS 17, visionOS 1, watchOS 10, *)
 #Preview {
-    Button("Pulsating") {
+    Button("Hello World!") {
 
     }
     .pulsating()