+++ Sources/SwiftExtras/Styles/GrayButtonStyle.swift
@@ -19,7 +19,7 @@ import SwiftUI
 ///
 /// Example:
 /// ```swift
-/// Button("Hello, World!") {}
+/// Button("Hello World!") {}
 ///     .buttonStyle(.gray)
 /// ```
 @available(iOS 17.0, macOS 14.0, *)
@@ -63,7 +63,7 @@ extension ButtonStyle where Self == GrayButtonStyle {
     ///
     /// Example:
     /// ```swift
-    /// Button("Hello, World!") {}
+    /// Button("Hello World!") {}
     ///     .buttonStyle(.gray)
     /// ```
     public static var gray: GrayButtonStyle { .init() }
@@ -73,10 +73,10 @@ extension ButtonStyle where Self == GrayButtonStyle {
 @available(iOS 17, macOS 14, tvOS 17, visionOS 1, watchOS 10, *)
 #Preview {
     VStack {
-        Button("Hello, World!") {}
+        Button("Hello World!") {}
             .buttonStyle(.gray)
 
-        Button("Hello, World!") {}
+        Button("Hello World!") {}
             .buttonStyle(.gray)
             .disabled(true)
     }